requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
banks==2.3.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecated==1.3.1
dirtyjson==1.0.8
fastapi==0.128.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
gitpython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inngest==0.5.15
jcs==0.2.1
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llama-index-core==0.14.13
llama-index-embeddings-gemini==0.4.2
llama-index-instrumentation==0.4.2
llama-index-llms-gemini==0.6.2
llama-index-readers-file==0.5.6
llama-index-workflows==2.13.1
markupsafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy-extensions==1.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
packaging==26.0
pandas==2.3.3
pillow==10.4.0
platformdirs==4.5.1
portalocker==3.2.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
pypdf==6.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
pyyaml==6.0.3
qdrant-client==1.16.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
setuptools==80.10.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
sqlalchemy==2.0.46
starlette==0.50.0
streamlit==1.53.1
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
wrapt==2.0.1
yarl==1.22.0
